Keep Tensormesh test diff within size gate
